Creando Modelos del Modulo de Avisos

+++ database/migrations/2021_11_16_224445_create_warning_sub_categories_table.php
@@ -16,7 +16,7 @@ class CreateWarningSubCategoriesTable extends Migration
         Schema::create('warning_sub_categories', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->unsignedSmallInteger('warning_category_id');
+            $table->unsignedBigInteger('warning_category_id');
             $table->timestamps();
 
             $table->foreign('warning_category_id')->references('id')->on('warning_categories')->onDelete('cascade');
